+++ packages/lab/src/banner/BannerCloseButton.tsx
@@ -1,0 +1,8 @@
+import { Button, ButtonProps } from "@salt-ds/core";
+import { CloseIcon } from "@salt-ds/icons";
+
+export const BannerCloseButton = (props: ButtonProps) => (
+  <Button {...props} variant="secondary">
+    <CloseIcon />
+  </Button>
+);