+++ packages/lab/src/form-field-next/FormFieldHelperText.tsx
@@ -1,0 +1,39 @@
+import clsx from "clsx";
+import { makePrefixer, StatusIndicator, Text, TextProps } from "@salt-ds/core";
+import { useFormFieldPropsNext } from "../form-field-context-next";
+
+import "./FormFieldHelperText.css";
+
+const withBaseName = makePrefixer("saltFormFieldHelperText");
+
+export const FormFieldHelperText = ({
+  className,
+  children,
+  ...restProps
+}: Omit<TextProps<"label">, "variant" | "styleAs">) => {
+  const { a11yProps, disabled, readOnly, validationStatus } =
+    useFormFieldPropsNext();
+
+  return (
+    <div
+      className={clsx(
+        withBaseName(),
+        { [withBaseName("withValidation")]: validationStatus },
+        className
+      )}
+    >
+      {!disabled && !readOnly && validationStatus && (
+        <StatusIndicator status={validationStatus} />
+      )}
+      <Text
+        disabled={disabled}
+        id={a11yProps?.["aria-describedby"]}
+        variant="secondary"
+        styleAs="label"
+        {...restProps}
+      >
+        {children}
+      </Text>
+    </div>
+  );
+};