+++ site/src/components/components/useLivePreviewControls.ts
@@ -1,0 +1,9 @@
+import { useContext } from "react";
+import {
+  LivePreviewContextType,
+  LivePreviewContext,
+} from "./LivePreviewControls";
+
+export const useLivePreviewControls = (): LivePreviewContextType => {
+  return useContext(LivePreviewContext);
+};